Minor edit

+++ cogs/error_handlers.py
@@ -35,18 +35,19 @@ class Error_Handlers:
                         "Missing Permissions"):
             return
         elif "The check functions for command" in str(exc):
-            await ctx.send("This command can only be used in NSFW channels", delete_after=7.0)
-            await asyncio.sleep(7)
-            try:
-                await ctx.message.delete()
-            except:
-                pass
-            return
+            if ctx.command.cog_name == "NSFW":
+                await ctx.send("This command can only be used in NSFW channels", delete_after=6.0)
+                await asyncio.sleep(6)
+                try:
+                    await ctx.message.delete()
+                except:
+                    pass
+                return
         elif str(exc) == "You do not own this bot.":
             app = await self.bot.application_info()
             await ctx.send(
-                f"Only my owner, **{app.owner.name}**, can use that command", delete_after=7.0)
-            await asyncio.sleep(7)
+                f"Only my owner, **{app.owner.name}**, can use that command", delete_after=6.0)
+            await asyncio.sleep(6)
             try:
                 await ctx.message.delete()
             except:
